User profile

+++ app/Http/Controllers/UserController.php
@@ -6,73 +6,62 @@ use App\Http\Requests\UserStoreRequest;
 use App\Http\Requests\UserUpdateRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Inertia\Inertia;
 
 class UserController extends Controller
 {
     /**
      * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
-    public function index(Request $request)
-    {
-        $users = User::all();
-
-        return view('user.index', compact('users'));
-    }
-
-    /**
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
-    public function create(Request $request)
-    {
-        return view('user.create');
-    }
-
-
-    /**
-     * @param \Illuminate\Http\Request $request
      * @param \App\Models\User $user
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Request $request, User $user)
-    {
-        return view('user.show', compact('user'));
-    }
-
-    /**
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Models\User $user
-     * @return \Illuminate\Http\Response
+     * @return \Inertia\Response
      */
     public function edit(Request $request, User $user)
     {
-        return view('user.edit', compact('user'));
+        if ($user->id == Auth::id() || Auth::user()->isSiteAdmin()) {
+            return Inertia::render('Profile', compact('user'));
+        }
+
+        abort(403);
     }
 
     /**
      * @param \App\Http\Requests\UserUpdateRequest $request
      * @param \App\Models\User $user
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(UserUpdateRequest $request, User $user)
     {
-        $user->update($request->validated());
+        $user->update($request->safe(['name', 'email', 'language']));
+
+        // Handle password changes
+        if ($newPassword = $request->safe(['password'])) {
+            $newPassword['password'] = Hash::make($newPassword['password']);
+            $user->update($newPassword);
+            // if the user updated their password, then re-auth them because their session will be invalidated.
+            Auth::login($user);
+        }
 
         $request->session()->flash('user.id', $user->id);
 
-        return redirect()->route('user.index');
+        return redirect()->back();
     }
 
     /**
      * @param \Illuminate\Http\Request $request
      * @param \App\Models\User $user
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Request $request, User $user)
     {
-        $user->delete();
+        if ($user->id == Auth::id() || Auth::user()->isSiteAdmin()) {
+            $user->delete();
+            Auth::logout();
 
-        return redirect()->route('user.index');
+            return redirect()->route('homepage');
+        }
+
+        abort(403);
     }
 }
